<==== Mercurial data ====>
Branch: x_VMS-3408_cloud_sync

+++ common/src/recording/stream_recorder.cpp
@@ -478,7 +478,7 @@ void QnStreamRecorder::writeData(const QnConstAbstractMediaDataPtr& md, int stre
         else
             avPkt.dts = dts;
         const QnCompressedVideoData* video = dynamic_cast<const QnCompressedVideoData*>(md.get());
-        if (video && (quint64)video->pts != AV_NOPTS_VALUE)
+        if (video && video->pts != AV_NOPTS_VALUE)
             avPkt.pts = av_rescale_q(video->pts-m_startDateTime, srcRate, stream->time_base) + (avPkt.dts-dts);
         else
             avPkt.pts = avPkt.dts;
